requirements.txt
setuptools=40.2.0
python-socketio==4.2.0
eventlet==0.21.0
flask==1.1.1
numpy==1.13.3
